requirements.txt
luigi==3.6.0
pandas==2.3.1
requests>=2.28.0
sqlalchemy<2.0.0
tornado==6.5
flask==3.1.1
flask-login==0.6.3
psycopg2-binary>=2.9.0
python-dotenv==1.1.1
dvc==3.61.0
dvc_s3==3.2.2
scikit-learn==1.7.1
mlflow==3.1.4